requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
PyJWT>=2.8.0

# Salesforce integration
simple-salesforce>=1.12.5

# CDC/Streaming (CometD client)
cometd>=0.1.0
aiohttp>=3.9.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# LLM integration (for routing and template suggestion)
openai>=1.3.0
anthropic>=0.7.0

# Utilities
pydantic>=2.4.0
python-dateutil>=2.8.2
pytz>=2023.3

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0

# Logging and monitoring
structlog>=23.1.0
